[77] Create Home module at data project
