+++ data/src/main/java/co/nimblehq/compose/crypto/data/service/ApiService.kt
@@ -6,15 +6,6 @@ import retrofit2.http.*
 
 @Suppress("LongParameterList")
 interface ApiService {
-    @GET("coins/markets")
-    suspend fun getCoins(
-        @Query("ids") coinIds: String?,
-        @Query("vs_currency") currency: String,
-        @Query("price_change_percentage") priceChangePercentage: String,
-        @Query("order") itemOrder: String,
-        @Query("per_page") itemPerPage: Int,
-        @Query("page") page: Int
-    ): Response<List<CoinItemResponse>>
 
     @GET("coins/{id}")
     suspend fun getCoin(
