+++ domain/src/main/java/co/nimblehq/compose/crypto/domain/repository/CoinRepository.kt
@@ -5,15 +5,6 @@ import kotlinx.coroutines.flow.Flow
 
 @Suppress("LongParameterList")
 interface CoinRepository {
-    fun getCoins(
-        coinIds: List<String>? = null,
-        currency: String,
-        priceChangePercentage: String,
-        itemOrder: String,
-        itemPerPage: Int,
-        page: Int
-    ): Flow<List<CoinItem>>
-
     fun getCoinDetail(
         coinId: String,
         localization: Boolean = false,
